Added frontend to registration (sort of)
Had to find a different way to do registration and login to be able to make it easier for frontend team to make things (html is enough, there is no need to use backend framework's features).

The registration window is the one Pablo made, although so far it's only an universal registration (aka, not for vendors nor for customers), and as a result I removed some of registration sections (which are not yet in the system anyway)

Also added the logo you can zoom into (which was made by Gustas) in the home page.